criando rotas CRUD

+++ src/routes/memories.ts
@@ -1,14 +1,19 @@
 //rotas voltadas a memorias
 import { FastifyInstance } from "fastify";
-import { prisma } from "../libs/prisma";
 import { z } from 'zod';
+import { prisma } from "../libs/prisma";
 
 
 export async function memoriesRoutes(app: FastifyInstance) {
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify()
+  })
   //listagem de memorias
   app.get('/memories', async (request) => {
     const memories = await prisma.memory.findMany({
-      
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: 'asc',
       },
@@ -37,13 +42,12 @@ export async function memoriesRoutes(app: FastifyInstance) {
       },
     })
 
-    // if (!memory.isPublic && memory.userId !== request.user.sub) {
-    //   return reply.status(401).send()
-    // }
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
     return memory
   })
-
 
   //criação de uma memoria
   app.post('/memories', async (request) => {
@@ -60,7 +64,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: '5b489aff-4a99-4ea2-bfa3-09020b5dff80',
+        userId: request.user.sub
       },
     })
 
@@ -83,17 +87,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-    // let memory = await prisma.memory.findUniqueOrThrow({
-    //   where: {
-    //     id,
-    //   },
-    // })
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
 
-    // if (memory.userId !== request.user.sub) {
-    //   return reply.status(401).send()
-    // }
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
-    let memory = await prisma.memory.update({
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -114,6 +118,16 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
 
     const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
     await prisma.memory.delete({
       where: {
